Extract Pharos::Configuration::Host::ResolvConf and ::Route into separate files
