+++ lib/pharos/configuration/network.rb
@@ -9,8 +9,8 @@ module Pharos
         attribute :trusted_subnets, Pharos::Types::Array.of(Pharos::Types::String)
         attribute :no_masq_local, Pharos::Types::Strict::Bool.default(false)
 
-        # @param routes [Array<Pharos::Configuration::Host::Routes>]
-        # @return [Array<Pharos::Configuration::Host::Routes>]
+        # @param routes [Array<Pharos::Configuration::Routes>]
+        # @return [Array<Pharos::Configuration::Routes>]
         def self.filter_host_routes(routes)
           routes.reject{ |route|
             route.dev == 'weave'
@@ -22,8 +22,8 @@ module Pharos
         attribute :ipip_mode, Pharos::Types::String.default('Always')
         attribute :nat_outgoing, Pharos::Types::Strict::Bool.default(true)
 
-        # @param routes [Array<Pharos::Configuration::Host::Routes>]
-        # @return [Array<Pharos::Configuration::Host::Routes>]
+        # @param routes [Array<Pharos::Configuration::Routes>]
+        # @return [Array<Pharos::Configuration::Routes>]
         def self.filter_host_routes(routes)
           routes.reject{ |route|
             route.proto == 'bird' || route.dev =~ /^cali/
@@ -35,8 +35,8 @@ module Pharos
         attribute :manifest_path, Pharos::Types::String
         attribute :options, Pharos::Types::Hash
 
-        # @param _routes [Array<Pharos::Configuration::Host::Routes>]
-        # @return [Array<Pharos::Configuration::Host::Routes>]
+        # @param _routes [Array<Pharos::Configuration::Routes>]
+        # @return [Array<Pharos::Configuration::Routes>]
         def self.filter_host_routes(_routes)
           # There's no way to validate routes for a custom CNI setup
           []
@@ -78,8 +78,8 @@ module Pharos
         (IPAddr.new(service_cidr) | 10).to_s
       end
 
-      # @param routes [Array<Pharos::Configuration::Host::Routes>]
-      # @return [Array<Pharos::Configuration::Host::Routes>]
+      # @param routes [Array<Pharos::Configuration::Routes>]
+      # @return [Array<Pharos::Configuration::Routes>]
       def filter_host_routes(routes)
         case provider
         when 'weave'
